elliptic-curve: move `MockCurve` to `dev::mock_curve`
This makes way to add more functionality to the `dev` module that isn't
related to `MockCurve`, which currently occupies the entire module.

The types are re-exported with a deprecation, for now, though maybe we
can get rid of that before a release.
